requirements.txt
# VBUS Installer Requirements
# Core dependencies
PyQt6>=6.4.0
psutil>=5.9.0
GPUtil>=1.4.0
py-cpuinfo>=9.0.0
wmi>=1.5.1
pywin32>=305

# For GPU/CUDA detection
pynvml>=11.5.0
nvidia-ml-py>=12.535.0

# For system analysis
pandas>=2.0.0
numpy>=1.24.0

# For AI inferencing in BUS
torch>=2.0.0
transformers>=4.30.0

# For configuration
pyyaml>=6.0
toml>=0.10.2

# For logging and monitoring
rich>=13.0.0
tqdm>=4.65.0

# For packaging
pyinstaller>=5.10.0
